data-ingestion/integrated-vectorization/requirements.txt
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
requests>=2.31.0
tqdm>=4.66.0
